Safari-611.1.10/LayoutTests/platform/wpe/fast/canvas/webgl/index-validation-with-subsequent-draws-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'ext.drawElementsInstancedANGLE')
